Changed to TracerOptions file parsing to simple JSON, instead of name/type/value triplet JSON

+++ Source/TracerCUDA/AOTracer.cu
@@ -210,7 +210,7 @@ void AOTracer::AskOptions()
 {
     // Generate Tracer Object
     VariableList list;
-    list.emplace(SAMPLE_NAME, OptionVariable(options.sampleCount));
+    list.emplace(SAMPLE_NAME, OptionVariable(static_cast<int64_t>(options.sampleCount)));
 
     if(callbacks) callbacks->SendCurrentOptions(::Options(std::move(list)));
 }
